+++ src/pages/KeyboardMouseButtonsPage.h
@@ -17,27 +17,24 @@
  */
 #pragma once
 
-#include <QPushButton>
-#include <QList>
+#include <QKeyEvent>
 
-#include "../../include/CharMapMode.h"
+#include "FaceButtonsPage.h"
 
 namespace OWC {
-    class CharMapWidget final: public QWidget {
+    class KeyboardMouseButtonsPage final: public FaceButtonsPage {
         Q_OBJECT
 
-    private:
-        QList<QPushButton *> charBtnList;
-
-        void addKeys(const std::map<int, std::string> &keyMap, QHBoxLayout *lyt);
+    protected:
+        void keyPressEvent(QKeyEvent *event) override;
 
     public:
-        explicit CharMapWidget(CharMapMode mode = CharMapMode::Keyboard);
+        KeyboardMouseButtonsPage();
 
-    private slots:
-        void onKeyClicked();
+    protected slots:
+        void onResetBtnClicked() override;
 
     signals:
-        void keyPressed(const QString &key);
+        void resetKeyboardMouseButtons();
     };
 }