+++ CHANGELOG.md
@@ -1,9 +1,12 @@
-## Future (continuous release)
+## 2.0
 
 - Fix Windows device detection
 - Fix invalid function error in Windows
 - Refactor back buttons UI for controller v1
 - Allow case insensitive keys in yaml profiles
+- Support controller v2 (win 5, mini 25)
+- Add mapping mode switch to home page (v2 only)
+- Show current controller emulation mode in home page (v2 only)
 
 ## 1.0